extract to test/benchmark.
